+++ tests/test.rs
@@ -1,7 +1,14 @@
+use std::mem;
+
 use stacked_errors::Error;
 
 trait VerifyCapable: Send + Sync {}
 impl VerifyCapable for Error {}
 
 #[test]
-fn tests() {}
+fn tests() {
+    // thanks to thin-vec
+    assert_eq!(mem::size_of::<Error>(), mem::size_of::<usize>());
+    assert_eq!(mem::size_of::<Option<Error>>(), mem::size_of::<usize>());
+    assert_eq!(mem::size_of::<Result<(), Error>>(), mem::size_of::<usize>());
+}