+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.3.0] - 07-05-2023
 ### Changes
  - Added  `Send + Sync` bounds to the `BoxedError` variant so that the whole `Error` struct is now `Send + Sync`
+ - Used `ThinVec` to reduce the size of `Result<(), Error>` to be the same as `usize`
 
 ## [0.2.0] - 06-06-2023
 ### Additions